// testservices/models/VehicleUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DeploymentStatus } from './DeploymentStatus';
export type VehicleUpdate = {
    vehicle_name?: (string | null);
    launch_flags?: (string | null);
    assigned_bundle_name?: (string | null);
    deployment_status?: (DeploymentStatus | null);
    installed_bundle_name?: (string | null);
    assigned_group_name?: (string | null);
};